+++ Code/scenes/gameScene.cpp
@@ -12,7 +12,7 @@ using namespace sf;
 void GameScene::Load() {
 	// load level
 	float tileSize = 40.0f;
-	ls::loadLevelFile("./Assets/level_1.txt", tileSize);
+	ls::loadLevelFile("Assets/level_1.txt", tileSize);
 	float off = Engine::getWindowSize().y - (ls::getHeight() * tileSize);
 	ls::setOffset(Vector2f(0.0f, off));
 	// make level physics
